AP-279 refactor PaymentAttributeUpdater
