+++ Import/PaymentAttributeUpdaterInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AdyenPayment\Import;
+
+interface PaymentAttributeUpdaterInterface
+{
+    public function setReadonlyOnAdyenTypePaymentAttribute(bool $readOnly);
+}
